requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.2
matplotlib==3.7.1
Pillow==9.3.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3